production-requirements.txt
django
djangorestframework
django-cors-headers
uwsgi